question-classification/requirements.txt
pandas==0.22.0
scikit_learn==0.19.1
sklearn-crfsuite==0.3.6
numpy==1.13.3